Refactor: Take advantage of polymorphism with SpecListener
